mgr/dashboard: tests: move helper classes to __init__.py
