+++ src/pybind/mgr/dashboard/tests/test_ganesha.py
@@ -5,7 +5,7 @@ import unittest
 
 from mock import MagicMock, Mock
 
-from .helper import KVStoreMockMixin
+from . import KVStoreMockMixin
 from .. import mgr
 from ..settings import Settings
 from ..services import ganesha
