+++ src/pybind/mgr/dashboard/tests/test_grafana.py
@@ -1,4 +1,4 @@
-from .helper import ControllerTestCase
+from . import ControllerTestCase
 from ..controllers.grafana import Grafana
 from .. import mgr
